fix(playground): improve code structure and readability
