+++ sites/playground/web/src/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from './use-mobile';
 export * from './use-input-message';
-export * from './use-framework-config';
